Fix signed/unsigned warning in GLCaps::print
git-svn-id: http://skia.googlecode.com/svn/trunk@2319 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ gpu/src/GrGpuGL.cpp
@@ -2247,7 +2247,7 @@ void GrGpuGL::GLCaps::print() const {
         "Apple",
     };
     GrPrintf("MSAA Type: %s\n", gMSFBOExtStr[fMSFBOType]);
-    for (int i = 0; i < GR_ARRAY_COUNT(fAASamples); ++i) {
+    for (int i = 0; i < (int)GR_ARRAY_COUNT(fAASamples); ++i) {
         GrPrintf("AA Level %d has %d samples\n", i, fAASamples[i]);
     }
     GrPrintf("Max FS Uniform Vectors: %d\n", fMaxFragmentUniformVectors);
